refactor(storage): Split memory.rs, keys/datastore.rs, and index/mod.rs into directory modules
Phase 2 of #330. Pure code movement with zero behavioral changes:
- backends/memory.rs → backends/memory/{mod,store,transaction,iterator,tests}.rs
- keys/datastore.rs → keys/datastore/{mod,data_store_key,index_key,misc,tests}.rs
- index/mod.rs → extracted index_type.rs and tests.rs
