+++ crates/storage/src/backends/memory/transaction.rs
@@ -1,0 +1,353 @@
+use async_trait::async_trait;
+use parking_lot::Mutex;
+use std::collections::{BTreeMap, HashSet};
+use std::sync::Arc;
+use tokio::sync::RwLock;
+
+use super::iterator::MemoryIterator;
+use crate::backends::shared::ConflictTracker;
+use crate::corekv::{
+    AsyncTxnCallback, Error, IterOptions, Iterator, Reader, Result, Txn, TxnCallback, Writer,
+};
+
+/// In-memory transaction with snapshot isolation and conflict detection.
+///
+/// Transactions maintain a snapshot of the store at creation time and track
+/// pending changes. On commit, write-write conflicts are detected using
+/// optimistic concurrency control.
+pub(crate) struct MemoryTxn {
+    /// Reference to the store's data
+    pub(crate) store: Arc<RwLock<BTreeMap<Vec<u8>, Vec<u8>>>>,
+
+    /// Conflict tracker for write-write conflict detection
+    pub(crate) conflict_tracker: Arc<ConflictTracker>,
+
+    /// Version at which this transaction's snapshot was taken
+    pub(crate) read_version: u64,
+
+    /// Snapshot of store at transaction start (for reads)
+    pub(crate) snapshot: BTreeMap<Vec<u8>, Vec<u8>>,
+
+    /// Pending changes (Some(value) = set, None = delete)
+    pub(crate) pending: Mutex<BTreeMap<Vec<u8>, Option<Vec<u8>>>>,
+
+    /// Whether this is a read-only transaction
+    pub(crate) readonly: bool,
+
+    /// Whether the transaction has been discarded
+    pub(crate) discarded: Mutex<bool>,
+
+    /// Whether the transaction has been committed
+    pub(crate) committed: Mutex<bool>,
+
+    /// Callbacks for successful commit
+    pub(crate) on_success: Mutex<Vec<TxnCallback>>,
+    pub(crate) on_success_async: Mutex<Vec<AsyncTxnCallback>>,
+
+    /// Callbacks for failed commit
+    pub(crate) on_error: Mutex<Vec<TxnCallback>>,
+    pub(crate) on_error_async: Mutex<Vec<AsyncTxnCallback>>,
+
+    /// Callbacks for discard
+    pub(crate) on_discard: Mutex<Vec<TxnCallback>>,
+    pub(crate) on_discard_async: Mutex<Vec<AsyncTxnCallback>>,
+}
+
+impl MemoryTxn {
+    /// Get a value, checking pending changes first, then snapshot.
+    fn get_internal(&self, key: &[u8]) -> Option<Vec<u8>> {
+        // Check pending changes first
+        let pending = self.pending.lock();
+        if let Some(pending_value) = pending.get(key) {
+            return pending_value.clone();
+        }
+
+        // Fall back to snapshot
+        self.snapshot.get(key).cloned()
+    }
+
+    /// Check if a key exists.
+    fn has_internal(&self, key: &[u8]) -> bool {
+        // Check pending changes first
+        let pending = self.pending.lock();
+        if let Some(pending_value) = pending.get(key) {
+            return pending_value.is_some();
+        }
+
+        // Fall back to snapshot
+        self.snapshot.contains_key(key)
+    }
+
+    /// Execute sync callbacks with panic protection.
+    ///
+    /// Each callback is wrapped in catch_unwind to ensure that a panic in one
+    /// callback doesn't prevent execution of subsequent callbacks.
+    fn execute_callbacks(callbacks: Vec<TxnCallback>) {
+        for (i, callback) in callbacks.into_iter().enumerate() {
+            let result = std::panic::catch_unwind(std::panic::AssertUnwindSafe(callback));
+            if let Err(panic_info) = result {
+                tracing::error!(
+                    callback_index = i,
+                    panic = ?panic_info,
+                    "Transaction callback panicked - continuing with remaining callbacks"
+                );
+            }
+        }
+    }
+
+    /// Execute async callbacks with panic protection.
+    ///
+    /// Each callback is executed sequentially with panic catching via FutureExt::catch_unwind.
+    async fn execute_async_callbacks(callbacks: Vec<AsyncTxnCallback>) {
+        use futures::FutureExt;
+
+        for (i, callback) in callbacks.into_iter().enumerate() {
+            let future = callback();
+            // Wrap the future in AssertUnwindSafe and catch panics
+            let result = std::panic::AssertUnwindSafe(future).catch_unwind().await;
+            if let Err(panic_info) = result {
+                tracing::error!(
+                    callback_index = i,
+                    panic = ?panic_info,
+                    "Async callback panicked during execution - continuing with remaining callbacks"
+                );
+            }
+        }
+    }
+}
+
+#[async_trait]
+impl Reader for MemoryTxn {
+    async fn get(&self, key: &[u8]) -> Result<Option<Vec<u8>>> {
+        if *self.discarded.lock() {
+            return Err(Error::DiscardedTxn);
+        }
+
+        if key.is_empty() {
+            return Err(Error::EmptyKey);
+        }
+
+        Ok(self.get_internal(key))
+    }
+
+    async fn has(&self, key: &[u8]) -> Result<bool> {
+        if *self.discarded.lock() {
+            return Err(Error::DiscardedTxn);
+        }
+
+        if key.is_empty() {
+            return Err(Error::EmptyKey);
+        }
+
+        Ok(self.has_internal(key))
+    }
+
+    async fn get_size(&self, key: &[u8]) -> Result<Option<usize>> {
+        if *self.discarded.lock() {
+            return Err(Error::DiscardedTxn);
+        }
+
+        if key.is_empty() {
+            return Err(Error::EmptyKey);
+        }
+
+        Ok(self.get_internal(key).map(|v| v.len()))
+    }
+
+    async fn iterator(&self, opts: IterOptions) -> Result<Box<dyn Iterator>> {
+        if *self.discarded.lock() {
+            return Err(Error::DiscardedTxn);
+        }
+
+        // Merge snapshot and pending changes
+        let mut merged = self.snapshot.clone();
+        let pending = self.pending.lock();
+        for (key, value) in pending.iter() {
+            match value {
+                Some(v) => {
+                    merged.insert(key.clone(), v.clone());
+                }
+                None => {
+                    merged.remove(key);
+                }
+            }
+        }
+
+        Ok(Box::new(MemoryIterator::new(merged, opts)?))
+    }
+}
+
+#[async_trait]
+impl Writer for MemoryTxn {
+    async fn set(&mut self, key: &[u8], value: &[u8]) -> Result<()> {
+        if *self.discarded.lock() {
+            return Err(Error::DiscardedTxn);
+        }
+
+        if self.readonly {
+            return Err(Error::ReadOnlyTxn);
+        }
+
+        if key.is_empty() {
+            return Err(Error::EmptyKey);
+        }
+
+        self.pending
+            .lock()
+            .insert(key.to_vec(), Some(value.to_vec()));
+        Ok(())
+    }
+
+    async fn delete(&mut self, key: &[u8]) -> Result<()> {
+        if *self.discarded.lock() {
+            return Err(Error::DiscardedTxn);
+        }
+
+        if self.readonly {
+            return Err(Error::ReadOnlyTxn);
+        }
+
+        if key.is_empty() {
+            return Err(Error::EmptyKey);
+        }
+
+        self.pending.lock().insert(key.to_vec(), None);
+        Ok(())
+    }
+}
+
+#[async_trait]
+impl Txn for MemoryTxn {
+    async fn commit(self: Box<Self>) -> Result<()> {
+        // Check discarded first (matches redb backend order)
+        if *self.discarded.lock() {
+            tracing::warn!("Attempted to commit a discarded transaction");
+            let on_error = std::mem::take(&mut *self.on_error.lock());
+            let on_error_async = std::mem::take(&mut *self.on_error_async.lock());
+            Self::execute_callbacks(on_error);
+            Self::execute_async_callbacks(on_error_async).await;
+            return Err(Error::DiscardedTxn);
+        }
+
+        // Check if already committed (defensive - ownership prevents this in normal usage)
+        if *self.committed.lock() {
+            tracing::warn!("Attempted to commit an already committed transaction");
+            return Err(Error::Other("Transaction already committed".into()));
+        }
+
+        // Clone pending changes before awaiting (can't hold MutexGuard across await)
+        let pending = self.pending.lock().clone();
+
+        // Check for write-write conflicts before applying
+        if !pending.is_empty() {
+            let write_set: HashSet<Vec<u8>> = pending.keys().cloned().collect();
+            if let Err(e) = self
+                .conflict_tracker
+                .check_and_record(self.read_version, write_set)
+            {
+                let on_error = std::mem::take(&mut *self.on_error.lock());
+                let on_error_async = std::mem::take(&mut *self.on_error_async.lock());
+                Self::execute_callbacks(on_error);
+                Self::execute_async_callbacks(on_error_async).await;
+                return Err(e);
+            }
+        }
+
+        // Mark as committed
+        *self.committed.lock() = true;
+
+        // Apply pending changes to store
+        if !pending.is_empty() {
+            let mut store = self.store.write().await;
+
+            for (key, value) in pending.iter() {
+                match value {
+                    Some(v) => {
+                        store.insert(key.clone(), v.clone());
+                    }
+                    None => {
+                        store.remove(key);
+                    }
+                }
+            }
+        }
+
+        // Execute success callbacks
+        let on_success = std::mem::take(&mut *self.on_success.lock());
+        let on_success_async = std::mem::take(&mut *self.on_success_async.lock());
+        Self::execute_callbacks(on_success);
+        Self::execute_async_callbacks(on_success_async).await;
+
+        Ok(())
+    }
+
+    fn discard(self: Box<Self>) {
+        *self.discarded.lock() = true;
+
+        // Execute sync discard callbacks
+        let on_discard = std::mem::take(&mut *self.on_discard.lock());
+        Self::execute_callbacks(on_discard);
+
+        // Handle async callbacks: spawn them in background with warning
+        let on_discard_async = std::mem::take(&mut *self.on_discard_async.lock());
+        if !on_discard_async.is_empty() {
+            let callback_count = on_discard_async.len();
+            tracing::warn!(
+                count = callback_count,
+                "Transaction has async discard callbacks. Spawning in background - they may not complete if process exits. Consider using commit() instead of discard() when async callbacks are registered."
+            );
+
+            // Spawn async callbacks in background with error tracking
+            // NOTE: These may not complete if the process exits before they finish
+            tokio::spawn(async move {
+                Self::execute_async_callbacks(on_discard_async).await;
+                tracing::debug!(count = callback_count, "Async discard callbacks completed");
+            });
+        }
+    }
+
+    fn on_success(&mut self, callback: TxnCallback) {
+        self.on_success.lock().push(callback);
+    }
+
+    fn on_success_async(&mut self, callback: AsyncTxnCallback) {
+        self.on_success_async.lock().push(callback);
+    }
+
+    fn on_error(&mut self, callback: TxnCallback) {
+        self.on_error.lock().push(callback);
+    }
+
+    fn on_error_async(&mut self, callback: AsyncTxnCallback) {
+        self.on_error_async.lock().push(callback);
+    }
+
+    fn on_discard(&mut self, callback: TxnCallback) {
+        self.on_discard.lock().push(callback);
+    }
+
+    fn on_discard_async(&mut self, callback: AsyncTxnCallback) {
+        self.on_discard_async.lock().push(callback);
+    }
+
+    fn as_any(&self) -> &dyn std::any::Any {
+        self
+    }
+
+    fn as_any_mut(&mut self) -> &mut dyn std::any::Any {
+        self
+    }
+
+    fn is_readonly(&self) -> bool {
+        self.readonly
+    }
+
+    fn callback_count(&self) -> usize {
+        self.on_success.lock().len()
+            + self.on_success_async.lock().len()
+            + self.on_error.lock().len()
+            + self.on_error_async.lock().len()
+            + self.on_discard.lock().len()
+            + self.on_discard_async.lock().len()
+    }
+}
